Move logic into components
